+++ frontend/src/pages/ManagerSpaceEdit/units/Board.styles.ts
